+++ src/main/java/ru/mgubin/tbot/cash/UserDataCache.java
@@ -12,24 +12,24 @@ import java.util.Map;
 @Component
 public class UserDataCache implements DataCache
 {
-    private Map<Integer, BotState> usersBotStates = new HashMap<>();
-    private Map<Integer, LikeState> usersLikeStates = new HashMap<>();
-    private Map<Integer, User> usersProfileData = new HashMap<>();
-    private Map<Integer, SearchProfile> usersSearchListData = new HashMap<>();
+    private Map<Long, BotState> usersBotStates = new HashMap<>();
+    private Map<Long, LikeState> usersLikeStates = new HashMap<>();
+    private Map<Long, User> usersProfileData = new HashMap<>();
+    private Map<Long, SearchProfile> usersSearchListData = new HashMap<>();
 
     @Override
-    public void setUsersCurrentBotState(int userId, BotState botState)
+    public void setUsersCurrentBotState(long userId, BotState botState)
     {
         usersBotStates.put(userId, botState);
     }
 
     @Override
-    public void setUsersCurrentLikeState(int userId, LikeState likeState)
+    public void setUsersCurrentLikeState(long userId, LikeState likeState)
     {
         usersLikeStates.put(userId, likeState);
     }
     @Override
-    public BotState getUsersCurrentBotState(int userId)
+    public BotState getUsersCurrentBotState(long userId)
     {
         BotState botState = usersBotStates.get(userId);
         if (botState == null)
@@ -40,7 +40,7 @@ public class UserDataCache implements DataCache
         return botState;
     }
     @Override
-    public LikeState getUsersCurrentLikeState(int userId)
+    public LikeState getUsersCurrentLikeState(long userId)
     {
         LikeState likeState = usersLikeStates.get(userId);
         if (likeState == null)
@@ -51,7 +51,7 @@ public class UserDataCache implements DataCache
         return likeState;
     }
     @Override
-    public User getUserProfileData(int userId)
+    public User getUserProfileData(long userId)
     {
         User userProfileData = usersProfileData.get(userId);
         if (userProfileData == null)
@@ -62,19 +62,19 @@ public class UserDataCache implements DataCache
     }
 
     @Override
-    public void saveUserProfileData(int userId, User userProfileData)
+    public void saveUserProfileData(long userId, User userProfileData)
     {
         usersProfileData.put(userId, userProfileData);
     }
 
     @Override
-    public void saveUserListData(int userId, SearchProfile listProfileData)
+    public void saveUserListData(long userId, SearchProfile listProfileData)
     {
         usersSearchListData.put(userId, listProfileData);
     }
 
     @Override
-    public SearchProfile getUserListData(int userId)
+    public SearchProfile getUserListData(long userId)
     {
         SearchProfile listUsers = usersSearchListData.get(userId);
         return listUsers;
