+++ src/main/java/ru/mgubin/tbot/command/AskBirthdayCommand.java
@@ -1,13 +1,12 @@
 package ru.mgubin.tbot.command;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Message;
 import ru.mgubin.tbot.cash.UserDataCache;
 import ru.mgubin.tbot.db.UserDB;
 import ru.mgubin.tbot.entity.User;
 import ru.mgubin.tbot.enums.BotState;
-import ru.mgubin.tbot.printer.PrintProfile;
+import ru.mgubin.tbot.service.PrintProfile;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
@@ -25,7 +24,7 @@ public class AskBirthdayCommand implements Command
     @Override
     public OutputParameters invoke(Message message)
     {
-        int userId = message.getFrom().getId().intValue();
+        long userId = message.getFrom().getId();
         OutputParameters outputParameters = new OutputParameters();
         PrintProfile profile = new PrintProfile();
         UserDB userDB = new UserDB();
